requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
multidict==6.3.2
openai==1.70.0
packaging==25.0
propcache==0.3.1
psycopg==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.2
whitenoise==6.9.0
yarl==1.19.0
joblib
pandas
scikit-learn